requirements.txt
# Google API dependencies for YouTube uploads
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Existing dependencies (if any)
requests==2.31.0